+++ Kernel/GenericInterface/Operation/Ticket/TicketUpdate.pm
@@ -109,7 +109,7 @@ if applicable the created ArticleID.
                 #     Diff => 10080, # Pending time in minutes
                 #},
             },
-            Article {                                                          # optional
+            Article => {                                                          # optional
                 ArticleTypeID                   => 123,                        # optional
                 ArticleType                     => 'some article type name',   # optional
                 SenderTypeID                    => 123,                        # optional
@@ -117,7 +117,7 @@ if applicable the created ArticleID.
                 AutoResponseType                => 'some auto response type',  # optional
                 From                            => 'some from string',         # optional
                 Subject                         => 'some subject',
-                Body                            => 'some body'
+                Body                            => 'some body',
 
                 ContentType                     => 'some content type',        # ContentType or MimeType and Charset is required
                 MimeType                        => 'some mime type',
@@ -1820,6 +1820,8 @@ sub _TicketUpdate {
         }
     }
 
+    my $UnlockOnAway = 1;
+
     # update Ticket->Owner
     if ( $Ticket->{Owner} || $Ticket->{OwnerID} ) {
         my $Success;
@@ -1829,6 +1831,7 @@ sub _TicketUpdate {
                 TicketID => $TicketID,
                 UserID   => $Param{UserID},
             );
+            $UnlockOnAway = 0;
         }
         elsif ( defined $Ticket->{OwnerID} && $Ticket->{OwnerID} ne $TicketData{OwnerID} )
         {
@@ -1837,6 +1840,7 @@ sub _TicketUpdate {
                 TicketID  => $TicketID,
                 UserID    => $Param{UserID},
             );
+            $UnlockOnAway = 0;
         }
         else {
 
@@ -1945,6 +1949,7 @@ sub _TicketUpdate {
             HistoryType    => $Article->{HistoryType},
             HistoryComment => $Article->{HistoryComment} || '%%',
             AutoResponseType => $Article->{AutoResponseType},
+            UnlockOnAway     => $UnlockOnAway,
             OrigHeader       => {
                 From    => $From,
                 To      => $To,